Commands and base type was separated from each other
